payment_identifier: factor out bip21 functions to bip21.py to break cyclic dependencies, parse bolt11 only once, store invoice internally instead of bolt11 string add is_onchain method to indicate if payment identifier can be paid onchain
